Finished the return of tip angle

+++ src/magmed_camera/include/magmed_camera/imageProcess.h
@@ -7,7 +7,7 @@ namespace magmed_camera
     {
     public:
         // getTipAngle函数：获取尖端偏转角，输入图像宽，高和数据存储指针，输出尖端偏转角度
-        double getTipAngle(unsigned short Height, unsigned short Width, unsigned char * pData);
+        float getTipAngle(unsigned short Height, unsigned short Width, unsigned char * pData);
     };
 }
